Rename bezier => bezier-spline
