+++ packages/turf-bezier-spline/test.js
@@ -4,7 +4,7 @@ import path from 'path';
 import load from 'load-json-file';
 import write from 'write-json-file';
 import { featureCollection } from '@turf/helpers';
-import bezier from '.';
+import bezierSpline from '.';
 
 const directories = {
     in: path.join(__dirname, 'test', 'in') + path.sep,
@@ -19,12 +19,12 @@ const fixtures = fs.readdirSync(directories.in).map(filename => {
     };
 });
 
-test('turf-bezier', t => {
+test('turf-bezier-spline', t => {
     fixtures.forEach(fixture => {
         const filename = fixture.filename;
         const name = fixture.name;
         const geojson = fixture.geojson;
-        const spline = colorize(bezier(geojson));
+        const spline = colorize(bezierSpline(geojson));
         const results = featureCollection([spline, geojson]);
 
         if (process.env.REGEN) write.sync(directories.out + filename, results);
