+++ packages/turf-bezier-spline/test/in/simple.geojson
